Moved the flatpickr dependency information to the migration notes (#175)
It turns out that besides installing flatpickr, which is discussed in the know issues, it is also necessary to add the style sheet to the styles array in angular.json for the datepicker to work properly. Currently, this note is underneath the new features header, but as it can break your date picker, it should be moved to the migration notes.