Fixed an error on the migration page and added extra information
The migration page mentioned that the dependency @auth0/angular-jwt required a version upgrade. But later on the page it is stated that it needs to be removed. I removed it from the Matching dependencies section

It is required that deepmerge-ts is also added as a dependency. A new section has been added for this.